category plot. item labels. api
