+++ plugin/jvm/src/main/java/com/twosigma/beaker/chart/serializer/CategoryBarsSerializer.java
@@ -66,6 +66,8 @@ public class CategoryBarsSerializer extends CategoryGraphicsSerializer<CategoryB
       jgen.writeObjectField("outline", categoryBars.getDrawOutline());
     }
 
+    jgen.writeObjectField("labelPosition", categoryBars.getLabelPosition());
+
     jgen.writeEndObject();
   }
 }
